feat: remove older TopBar in favor of TopBarV2
