+++ components/Header.js
@@ -9,7 +9,6 @@ import { truncate } from '../lib/utils';
 
 import GlobalWarnings from './GlobalWarnings';
 import TopBar from './TopBar';
-import TopBarV2 from './TopBarV2';
 
 const messages = defineMessages({
   defaultTitle: {
@@ -99,7 +98,7 @@ class Header extends React.Component {
   }
 
   render() {
-    const { css, className, canonicalURL, withTopBar, LoggedInUser } = this.props;
+    const { css, canonicalURL, withTopBar } = this.props;
     return (
       <header>
         <Head>
@@ -120,21 +119,13 @@ class Header extends React.Component {
           {canonicalURL && <link rel="canonical" href={canonicalURL} />}
         </Head>
         <div id="top" />
-        {withTopBar &&
-          (LoggedInUser?.collective?.settings?.useNewTopBar ? (
-            <TopBarV2
-              showSearch={this.props.showSearch}
-              menuItems={this.props.menuItemsV2}
-              showProfileAndChangelogMenu={this.props.showProfileAndChangelogMenu}
-            />
-          ) : (
-            <TopBar
-              className={className}
-              showSearch={this.props.showSearch}
-              menuItems={this.props.menuItems}
-              showProfileAndChangelogMenu={this.props.showProfileAndChangelogMenu}
-            />
-          ))}
+        {withTopBar && (
+          <TopBar
+            showSearch={this.props.showSearch}
+            menuItems={this.props.menuItems}
+            showProfileAndChangelogMenu={this.props.showProfileAndChangelogMenu}
+          />
+        )}
         <GlobalWarnings collective={this.props.collective} />
       </header>
     );
