+++ components/TopBarMobileMenu.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { ChevronDown } from '@styled-icons/boxicons-regular/ChevronDown';
 import { Github } from '@styled-icons/fa-brands/Github';
 import { Slack } from '@styled-icons/fa-brands/Slack';
 import { Twitter } from '@styled-icons/fa-brands/Twitter';
@@ -9,7 +10,7 @@ import { FormattedMessage, injectIntl } from 'react-intl';
 import styled from 'styled-components';
 
 import Container from './Container';
-import { Box } from './Grid';
+import { Box, Flex } from './Grid';
 import Link from './Link';
 import StyledLink from './StyledLink';
 import StyledRoundButton from './StyledRoundButton';
@@ -23,15 +24,24 @@ const ListItem = styled.li`
   font-size: 15px;
   line-height: 18px;
   padding-top: 10px;
-  padding-bottom: 10px;
+  cursor: pointer;
   a:not(:hover) {
     color: #313233;
   }
 `;
 
+const SubListItem = styled(ListItem)`
+  padding-bottom: 10px;
+`;
+
 class TopBarMobileMenu extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { viewSolutionsMenu: false, viewProductsMenu: false, viewCompanyMenu: false };
+  }
+
   render() {
-    const { showMobileMenu, menuItems, closeMenu } = this.props;
+    const { showMobileMenu, closeMenu } = this.props;
 
     if (!showMobileMenu) {
       return null;
@@ -42,43 +52,103 @@ class TopBarMobileMenu extends React.Component {
         bg="white.full"
         width="100%"
         position="absolute"
-        right={[0, 16]}
-        top={[69, 75]}
+        right={[0, 0, 16]}
+        top={[69, 69, 75]}
         p={3}
         zIndex={3000}
         borderRadius="0px 0px 16px 16px"
         boxShadow="0px 8px 12px rgba(20, 20, 20, 0.16)"
         data-cy="user-menu"
       >
-        <Box as="ul" my={2} pl={0}>
-          {menuItems.discover && (
-            <ListItem>
-              <Link href="/search" onClick={closeMenu}>
-                <FormattedMessage id="menu.discover" defaultMessage="Discover" />
-              </Link>
-            </ListItem>
-          )}
-          {menuItems.howItWorks && (
-            <ListItem>
-              <Link href="/how-it-works" onClick={closeMenu}>
-                <FormattedMessage id="menu.howItWorks" defaultMessage="How it Works" />
-              </Link>
-            </ListItem>
-          )}
-          {menuItems.pricing && (
-            <ListItem>
-              <Link href="/pricing" onClick={closeMenu}>
-                <FormattedMessage id="menu.pricing" defaultMessage="Pricing" />
-              </Link>
-            </ListItem>
-          )}
-          {menuItems.docs && (
-            <ListItem>
-              <a href="/help" onClick={closeMenu}>
-                <FormattedMessage id="menu.docs" defaultMessage="Docs & Help" />
-              </a>
-            </ListItem>
-          )}
+        <Box as="ul" my={2} pl={0} pb="36px">
+          <ListItem>
+            <Flex
+              justifyContent="space-between"
+              onClick={() => this.setState(({ viewSolutionsMenu }) => ({ viewSolutionsMenu: !viewSolutionsMenu }))}
+            >
+              <FormattedMessage defaultMessage="Solutions" />
+              <ChevronDown size={20} />
+            </Flex>
+            {this.state.viewSolutionsMenu && (
+              <Box as="ul" my={2} pl="12px">
+                {/* TODO: Add this part back when the /collectives page is designed*/}
+                {/* <ListItem>*/}
+                {/*  <Link href="/collectives" onClick={closeMenu}>*/}
+                {/*    <FormattedMessage id="pricing.forCollective" defaultMessage="For Collectives" />*/}
+                {/*  </Link>*/}
+                {/* </ListItem>*/}
+                <SubListItem>
+                  <Link href="/become-a-sponsor" onClick={closeMenu}>
+                    <FormattedMessage defaultMessage="For Contributors" />
+                  </Link>
+                </SubListItem>
+                <SubListItem>
+                  <Link href="/become-a-host" onClick={closeMenu}>
+                    <FormattedMessage id="pricing.fiscalHost" defaultMessage="For Fiscal Hosts" />
+                  </Link>
+                </SubListItem>
+              </Box>
+            )}
+          </ListItem>
+          <hr />
+          <ListItem>
+            <Flex
+              justifyContent="space-between"
+              onClick={() => this.setState(({ viewProductsMenu }) => ({ viewProductsMenu: !viewProductsMenu }))}
+            >
+              <FormattedMessage id="ContributionType.Product" defaultMessage="Product" />
+              <ChevronDown size={20} />
+            </Flex>
+            {this.state.viewProductsMenu && (
+              <Box as="ul" my={2} pl="12px">
+                <SubListItem>
+                  <Link href="/pricing" onClick={closeMenu}>
+                    <FormattedMessage id="menu.pricing" defaultMessage="Pricing" />
+                  </Link>
+                </SubListItem>
+                <SubListItem>
+                  <Link href="/how-it-works" onClick={closeMenu}>
+                    <FormattedMessage id="menu.howItWorks" defaultMessage="How it Works" />
+                  </Link>
+                </SubListItem>
+                <SubListItem>
+                  <Link href="/fiscal-hosting" onClick={closeMenu}>
+                    <FormattedMessage id="editCollective.fiscalHosting" defaultMessage="Fiscal Hosting" />
+                  </Link>
+                </SubListItem>
+              </Box>
+            )}
+          </ListItem>
+          <hr />
+          <ListItem>
+            <Flex
+              justifyContent="space-between"
+              onClick={() => this.setState(({ viewCompanyMenu }) => ({ viewCompanyMenu: !viewCompanyMenu }))}
+            >
+              <FormattedMessage id="company" defaultMessage="Company" />
+              <ChevronDown size={20} />
+            </Flex>
+            {this.state.viewCompanyMenu && (
+              <Box as="ul" my={2} pl="12px">
+                <SubListItem>
+                  <a href="https://blog.opencollective.com/" onClick={closeMenu}>
+                    <FormattedMessage id="company.blog" defaultMessage="Blog" />
+                  </a>
+                </SubListItem>
+                <SubListItem>
+                  <Link href="/e2c" onClick={closeMenu}>
+                    <FormattedMessage id="OC.e2c" defaultMessage="Exit to Community" />
+                  </Link>
+                </SubListItem>
+              </Box>
+            )}
+          </ListItem>
+          <hr />
+          <ListItem>
+            <Link href="/help" onClick={closeMenu}>
+              <FormattedMessage defaultMessage="Help & Support" />
+            </Link>
+          </ListItem>
         </Box>
         <Container
           display="flex"
@@ -88,29 +158,30 @@ class TopBarMobileMenu extends React.Component {
           order={['2', null, '3']}
           borderRadius={16}
           background="#F7F8FA"
+          justifyContent="space-between"
         >
           <StyledLink href="https://blog.opencollective.com/" openInNewTab onClick={closeMenu}>
-            <StyledRoundButton size={40} mr={2}>
+            <StyledRoundButton size={40}>
               <Blog size={17} color="#9D9FA3" />
             </StyledRoundButton>
           </StyledLink>
           <StyledLink href="https://twitter.com/opencollect" openInNewTab onClick={closeMenu}>
-            <StyledRoundButton size={40} mr={2}>
+            <StyledRoundButton size={40}>
               <Twitter size={17} color="#9D9FA3" />
             </StyledRoundButton>
           </StyledLink>
           <StyledLink href="https://github.com/opencollective" openInNewTab onClick={closeMenu}>
-            <StyledRoundButton size={40} mr={2}>
+            <StyledRoundButton size={40}>
               <Github size={17} color="#9D9FA3" />
             </StyledRoundButton>
           </StyledLink>
           <StyledLink href="https://slack.opencollective.com" openInNewTab onClick={closeMenu}>
-            <StyledRoundButton size={40} mr={2}>
+            <StyledRoundButton size={40}>
               <Slack size={17} color="#9D9FA3" />
             </StyledRoundButton>
           </StyledLink>
-          <StyledLink as={Link} href="/contact" openInNewTab onClick={closeMenu}>
-            <StyledRoundButton size={40} mr={2}>
+          <StyledLink href="mailto:info@opencollective.com" openInNewTab onClick={closeMenu}>
+            <StyledRoundButton size={40}>
               <Mail size={19} color="#9D9FA3" />
             </StyledRoundButton>
           </StyledLink>
@@ -122,7 +193,6 @@ class TopBarMobileMenu extends React.Component {
 
 TopBarMobileMenu.propTypes = {
   showMobileMenu: PropTypes.bool,
-  menuItems: PropTypes.object,
   closeMenu: PropTypes.func,
 };
