+++ pages/contribution-flow.js
@@ -206,7 +206,7 @@ class NewContributionFlowPage extends React.Component {
       <Page
         {...this.getPageMetadata()}
         showFooter={false}
-        menuItemsV2={{ solutions: false, product: false, company: false, docs: false }}
+        menuItems={{ solutions: false, product: false, company: false, docs: false }}
         showSearch={false}
       >
         {this.renderPageContent()}
