+++ pages/help-and-support.js
@@ -16,8 +16,6 @@ import Page from '../components/Page';
 import StyledButton from '../components/StyledButton';
 import StyledLink from '../components/StyledLink';
 
-const menuItems = { pricing: true };
-
 const messages = defineMessages({
   defaultTitle: {
     id: 'OC.helpAndSupport',
@@ -74,7 +72,7 @@ const HelpAndSupport = ({ action, formConfirmation }) => {
   const { formatMessage } = useIntl();
 
   return (
-    <Page menuItems={menuItems} description={formatMessage(messages.defaultTitle)}>
+    <Page description={formatMessage(messages.defaultTitle)}>
       {action === 'contact' ? (
         renderFormContent(formConfirmation)
       ) : (
