+++ pages/marketingPage.js
@@ -33,10 +33,6 @@ function importAll(r) {
   return map;
 }
 
-function getmenuItem() {
-  return { pricing: false, howItWorks: false };
-}
-
 class MarketingPage extends React.Component {
   static getInitialProps({ query: { pageSlug } }) {
     return { pageSlug };
@@ -85,7 +81,7 @@ class MarketingPage extends React.Component {
     return (
       <Fragment>
         <div>
-          <Header LoggedInUser={LoggedInUser} menuItems={getmenuItem()} />
+          <Header LoggedInUser={LoggedInUser} />
           <Body>
             <style type="text/css" dangerouslySetInnerHTML={{ __html: style }} />
             <div className={className} dangerouslySetInnerHTML={{ __html: html }} />
