--- stories/layout/TopBarV2.stories.mdx
@@ -1,65 +1,0 @@
-import { ArgsTable, Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
-import { cloneDeep, update } from 'lodash';
-import TopBarV2 from '../../components/TopBarV2';
-import { UserContext } from '../../components/UserProvider';
-import theme from '../../lib/theme';
-import { loggedInUser, newLoggedInUser } from '../mocks/loggedInUser';
-
-<Meta
-  title="Layout/TopBarV2"
-  component={TopBarV2}
-  parameters={{
-    actions: {
-      handles: ['mouseover', 'click'],
-    },
-  }}
-/>
-
-# TopBarV2
-
-## Logged out
-
-export const LoggedOutStory = props => (
-  <UserContext.Provider value={{ LoggedInUser: null, loadingLoggedInUser: false }}>
-    <TopBarV2 {...props} />
-  </UserContext.Provider>
-);
-
-<Story name="Logged out">{LoggedOutStory.bind({})}</Story>
-
-## Logged in (new user)
-
-export const LoggedInStory = props => (
-  <UserContext.Provider value={{ LoggedInUser: newLoggedInUser, loadingLoggedInUser: false }}>
-    <TopBarV2 {...props} />
-  </UserContext.Provider>
-);
-
-<Story name="Logged in (new user)">{LoggedInStory.bind({})}</Story>
-
-## Logged in (simple collective admin)
-
-export const LoggedInCollectiveAdmin = props => (
-  <UserContext.Provider
-    value={{
-      loadingLoggedInUser: false,
-      LoggedInUser: update(cloneDeep(loggedInUser), 'memberOf', memberOf =>
-        memberOf.filter(m => m.role === 'ADMIN' && m.collective.type === 'COLLECTIVE').slice(0, 3),
-      ),
-    }}
-  >
-    <TopBarV2 {...props} />
-  </UserContext.Provider>
-);
-
-<Story name="Logged in (collective admin)">{LoggedInCollectiveAdmin.bind({})}</Story>
-
-## Logged in (power user)
-
-export const LoggedInPowerUserStory = props => (
-  <UserContext.Provider value={{ LoggedInUser: loggedInUser, loadingLoggedInUser: false }}>
-    <TopBarV2 {...props} />
-  </UserContext.Provider>
-);
-
-<Story name="Logged in (power user)">{LoggedInPowerUserStory.bind({})}</Story>
